Name changed to official name TecaliApp
