+++ android/app/BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.ayuntamientoapp",
+    package = "com.tecaliapp",
 )
 
 android_resource(
     name = "res",
-    package = "com.ayuntamientoapp",
+    package = "com.tecaliapp",
     res = "src/main/res",
 )
