+++ android/app/src/main/java/com/tecaliapp/MainActivity.java
@@ -1,0 +1,35 @@
+package com.tecaliapp;
+
+import com.facebook.react.ReactActivity;
+import com.facebook.react.ReactActivityDelegate;
+import com.facebook.react.ReactRootView;
+import android.content.Intent;
+
+public class MainActivity extends ReactActivity {
+
+    /**
+     * Returns the name of the main component registered from JavaScript.
+     * This is used to schedule rendering of the component.
+     */
+    @Override
+    protected String getMainComponentName() {
+        return "TecaliApp";
+    }
+
+    // @Override
+    // protected ReactActivityDelegate createReactActivityDelegate() {
+    //   return new ReactActivityDelegate(this, getMainComponentName()) {
+    //     @Override
+    //     protected ReactRootView createRootView() {
+    //      return new RNGestureHandlerEnabledRootView(MainActivity.this);
+    //     }
+    //   };
+    // }
+
+    // Intent react-native-fcm
+    @Override
+    public void onNewIntent(Intent intent) {
+        super.onNewIntent(intent);
+        setIntent(intent);
+    }
+}
