+++ android/settings.gradle
@@ -1,4 +1,4 @@
-rootProject.name = 'AyuntamientoApp'
+rootProject.name = 'TecaliApp'
 include ':react-native-add-calendar-event'
 project(':react-native-add-calendar-event').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-add-calendar-event/android')
 include ':react-native-gesture-handler'
