+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "AyuntamientoApp",
-  "displayName": "AyuntamientoApp"
+  "name": "TecaliApp",
+  "displayName": "TecaliApp"
 }
