+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "ayuntamiento-app",
+  "name": "tecaliapp",
   "version": "0.0.1",
   "private": true,
   "scripts": {
